+++ Maui.RevenueCat.InAppBilling/Enums/IntroElegibilityStatus.cs
@@ -1,6 +1,3 @@
-#if IOS
-using Maui.RevenueCat.iOS;
-#endif
 namespace Maui.RevenueCat.InAppBilling.Enums;
 
 public enum IntroElegibilityStatus
@@ -10,24 +7,3 @@ public enum IntroElegibilityStatus
     NoIntroOfferExists,
     Unknown
 }
-
-#if IOS
-public static class Extensions
-{
-    public static IntroElegibilityStatus Convert(this RCIntroEligibilityStatus eligibility)
-    {
-        switch (eligibility)
-        {
-            case RCIntroEligibilityStatus.Ineligible:
-                return IntroElegibilityStatus.Ineligible;
-            case RCIntroEligibilityStatus.Eligible:
-                return IntroElegibilityStatus.Eligible;
-            case RCIntroEligibilityStatus.NoIntroOfferExists:
-                return IntroElegibilityStatus.NoIntroOfferExists;
-            case RCIntroEligibilityStatus.Unknown:
-            default:
-                return IntroElegibilityStatus.Unknown;
-        }
-    }
-}
-#endif
